expectDifference/expectNoDifference rename cleanups
While upgrading I noticed some of the docs / existing tests still use the old naming, so took a few moments to update.
