+++ Tests/CustomDumpTests/ExpectDifferenceTests.swift
@@ -2,31 +2,31 @@ import CustomDump
 import XCTest
 
 @available(*, deprecated)
-class XCTAssertDifferencesTests: XCTestCase {
-  func testXCTAssertDifference() {
+class ExpectDifferencesTests: XCTestCase {
+  func testExpectDifference() {
     var user = User(id: 42, name: "Blob")
     func increment<Value>(_ root: inout Value, at keyPath: WritableKeyPath<Value, Int>) {
       root[keyPath: keyPath] += 1
     }
 
-    XCTAssertDifference(user) {
+    expectDifference(user) {
       increment(&user, at: \.id)
     } changes: {
       $0.id = 43
     }
   }
 
-  func testXCTAssertDifference_NonExhaustive() {
+  func testExpectDifference_NonExhaustive() {
     let user = User(id: 42, name: "Blob")
 
-    XCTAssertDifference(user) {
+    expectDifference(user) {
       $0.id = 42
       $0.name = "Blob"
     }
   }
 
   #if DEBUG && compiler(>=5.4) && (os(iOS) || os(macOS) || os(tvOS) || os(watchOS))
-    func testXCTAssertDifference_Failure() {
+    func testExpectDifference_Failure() {
       var user = User(id: 42, name: "Blob")
       func increment<Value>(_ root: inout Value, at keyPath: WritableKeyPath<Value, Int>) {
         root[keyPath: keyPath] += 1
@@ -34,7 +34,7 @@ class XCTAssertDifferencesTests: XCTestCase {
 
       XCTExpectFailure()
 
-      XCTAssertDifference(user) {
+      expectDifference(user) {
         increment(&user, at: \.id)
       } changes: {
         $0.id = 44
